apps/scoring-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
httpx==0.25.2
google-generativeai==0.3.2
